Fix/update `isPrimary` and `canBePrimary` logic, drop 'Use for sending' option for API keys
`canBePrimary` is now always false for API keys.
While using a non-primary key for sending is allowed in principle, this could interfere
with e.g. automatic forwarding, hence we drop the feature altogether for internal contacts.
Note that the existing logic was already broken, and selecting the 'Use for sending'
button had no practical effect.
In the context of WKD, at most one key would be returned by the server, hence
the feature didn't apply.
The option remains supported for uploaded keys.

`isPrimary` had inconsistent behaviour for external contacts, depending on
whether the key was uploaded or from WKD. In the first case, the label would
only be shown with encryption enabled. In the past, encryption was enforced
with WKD keys, but since we implemented the option to disable encryption to them
as well, the display logic required updating too.