CryptoAPI: switch to OpenPGP.js v6 types
This is primarily an internal change, that does not affect the apps
(except for the changing `AlgorithmInfo` and `enums.curve` types).
An adapter is introduced to make the (stable) OpenPGP.js v5 types compatible
with the v6 canary ones (replacing the existing adapter which was used to convert
v6 types to v5).

This commit does not introduce support for v6 features.

CRYPTO-199