CryptoProxy: drop monitoring for key compatibility errors
No errors collected, which means we can go ahead and check by default on all imports
(this will be implemented separately, as the checks need to depend on whether the user
supports v6 keys or not).
We can drop the monitoring in the meantime, as we don't want to get reports for v6 keys.