Fix SKL parsing with colliding key fingerprint
Some old users have duplicate address keys, meaning the
fingerprint of the primary key packet and all subkeys are the same.
As a result, the SHA256Fingerprints entries in the SKL will also
be identical, but the keys should still be treated as separate entities
(in particular, keeping track of differing 'Primary' flag values).
The problem is that if duplicate keys were present, with one set as primary,
the extracted active address key would not be marked as primary,
triggering errors about missing primary key in e.g. `getPrimaryAddressKeysForSigning`.

The existing code would also mishandle keys with the same primary key packet
and different subkeys.