Mail: rename custom VERIFICATION_STATUS to MAIL_VERIFICATION_STATUS
To better differentiate it from the VERIFICATION_STATUS exported by the crypto package.
MAIL_VERIFICATION_STATUS is effectively an extension of that, but the definitions are not bound together,
due to TS limitations with enum types.
Renaming the Mail's enum makes TS enforce proper type checks.
Hence, to facilitate conversion between the two, a new `getMailVerificationStatus` helper has been introduced.